+++ src/Oauth1.php
@@ -93,6 +93,16 @@ class Oauth1
         }
     }
 
+    public function __serialize(): array
+    {
+        throw new \LogicException(self::class.' should never be serialized');
+    }
+
+    public function __unserialize(array $data): void
+    {
+        throw new \LogicException(self::class.' should never be unserialized');
+    }
+
     /**
      * Called when the middleware is handled.
      *